requirement.txt
absl-py==0.6.0
astor==0.7.1
bleach==1.5.0
boto==2.49.0
boto3==1.9.31
botocore==1.12.31
bz2file==0.98
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
gast==0.2.0
gensim==3.6.0
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
grpcio==1.16.0
gspread==3.0.1
h5py==2.8.0
html5lib==0.9999999
httplib2==0.12.0
idna==2.7
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lxml==4.2.5
Markdown==3.0.1
nltk==3.3
numpy==1.15.3
oauth2client==4.1.3
pandas==0.23.4
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyperclip==1.7.0
python-dateutil==2.7.3
pytz==2018.7
PyYAML==3.13
requests==2.20.0
rsa==4.0
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.24
Werkzeug==0.14.1
